requirements.txt
# CineSense Movie Recommendation System
# Requirements File

# Django Framework
Django>=4.2,<5.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Data Visualization
matplotlib>=3.7.0

# Machine Learning
scikit-learn>=1.3.0

# HTTP Requests for External APIs
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Optional: For production deployment
gunicorn>=21.0.0

# Optional: For better SQLite support
# For production, consider:
# psycopg2-binary>=2.9.0  # PostgreSQL

# Development tools (optional)
# django-debug-toolbar>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
